build: split CW FHIR proxy
Ref: metriport/metriport-internal#1040
