+++ packages/api/src/command/medical/admin/upload-doc.ts
@@ -1,14 +1,14 @@
 import { DocumentReference } from "@medplum/fhirtypes";
 import { FileData } from "@metriport/core/external/aws/lambda-logic/document-uploader";
+import { randomInt } from "@metriport/shared/common/numbers";
 import dayjs from "dayjs";
+import { cloneDeep } from "lodash";
 import { makeFhirApi } from "../../../external/fhir/api/api-factory";
 import { createDocReferenceContent, getFHIRDocRef } from "../../../external/fhir/document";
 import { metriportDataSourceExtension } from "../../../external/fhir/shared/extensions/metriport";
 import { Config } from "../../../shared/config";
 import { capture } from "../../../shared/notifications";
-import { randomInt } from "@metriport/shared/common/numbers";
 import { getPatientOrFail } from "../patient/get-patient";
-import { cloneDeep } from "lodash";
 
 const apiUrl = Config.getApiUrl();
 const docContributionUrl = `${apiUrl}/doc-contribution/commonwell/`;
