+++ packages/api/src/external/commonwell/proxy/get-org-or-fail.ts
@@ -1,6 +1,6 @@
 import { Organization } from "@metriport/core/domain/organization";
-import NotFoundError from "../../errors/not-found";
-import { OrganizationModel } from "../../models/medical/organization";
+import NotFoundError from "../../../errors/not-found";
+import { OrganizationModel } from "../../../models/medical/organization";
 
 // Didn't reuse getOrganizationOrFail bc we don't have `cxId` in this context and
 // we want to keep that function requiring `cxId` to avoid cross-tenant data access
